refactor(KERN-1560): split extract-where.ts into focused leaf modules
extract-filters/ now:
- extract-where.ts (155 lines): orchestrator + OR/AND logic
- extract-comparison.ts (100 lines): =, !=, >, <, >=, <=, NULL handling
- extract-between.ts (55 lines): BETWEEN (numeric + date range)
- extract-operators.ts (90 lines): IN/NOT IN, IS NULL, LIKE/ILIKE
- extract-having.ts (95 lines): HAVING extraction
- filter-schema.ts (90 lines): type inference + dimension schema

work-item: KERN-1560
